Add gpg verification of rpm source on Fedora
Fedora Packaging Guidelines state that packages should verify OpenPGP
signatures (if available) as part of the RPM build process.  This must
be done in the %prep section before any potentially compromised code
can be executed.

The gnupg2 package is always listed as a BuildRequires (instead
of placing it in an "%if %{with verify}" block as a conditional
dependency).  Consequently, the "%bcond_without verify" can be changed
to %bcond_with (i.e., disabling verification by default) while still
allowing verification by adding "--with verify".

"rpmbuild -tb" looks inside the tarball for only the rpm spec file.
Source verification requires the keyring and signature to also reside
in the directory:

  $ ls
  dun.gpg  munge-0.5.13.tar.xz  munge-0.5.13.tar.xz.asc
  $ rpmbuild -tb munge-0.5.13.tar.xz

Source verification can be disabled if the keyring and signature are
not present:

  $ ls
  munge-0.5.13.tar.xz
  $ rpmbuild -tb --without verify munge-0.5.13.tar.xz

Since the %{gpgverify} RPM macro is not defined on CentOS, builds
now fail there unless "--without verify" is specified.

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_source_file_verification
- https://rpm.org/user_doc/conditional_builds.html

Tested:
- CentOS Stream, 8.0.1905, 7.7.1908 (--without verify)
- Fedora 31, 30

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>