Sharness: Remove tests to/from invalid files
On FreeBSD (12.1, 11.4, 11.3) and NetBSD (9.0, 8.1, 7.2), the following
test fails when run with "root=/tmp/munge-test-$$":

  0012-munge-cmdline.t 24 - munge --input from invalid file

This test attempts to read data for a credential payload from the file
"." -- i.e., a directory, and not a regular file.  It is expected
to fail, and on most platforms it does.  However, it unexpectedly
succeeds if the input file is on a FreeBSD ufs or NetBSD ffs filesystem
(where it uses the directory file contents as the payload data),
but fails if the input file is on an nfs or tmpfs filesystem on
those platforms.  Note that this test fails as expected on OpenBSD
ffs and nfs filesystems.

This passed testing for 0.5.14 because the test suite ran in an
nfs directory.  But recent testing with "root=/tmp/munge-test-$$"
uncovered the failure since the "root" variable moved the input file
to a different filesystem.

Since the munge and unmunge client executables do not explicitly
check whether the input or output files are regular files, remove the
sharness checks that test for an expected failure when specifying an
invalid input, metadata, or output file.

Tested:
- Arch Linux
- CentOS Stream, 8.2.2004, 7.8.2003, 6.10
- Debian sid, 10.4, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 32, 31, 30
- FreeBSD 12.1, 11.4, 11.3
- NetBSD 9.0, 8.1, 7.2
- OpenBSD 6.7, 6.6, 6.5
- openSUSE 15.2, 15.1
- Raspberry Pi OS
- Ubuntu 20.04.1 LTS, 18.04.4 LTS, 16.04.6 LTS, 14.04.6 LTS, 12.04.5 LTS