Sharness: Replace MUNGE_ROOT with root
The "root" variable specifies a directory to be prepended to the
sharness "trash" directories.  A trash directory is created for each
test script to store temporary data, and that directory becomes the
HOME directory for the given test.

The "MUNGE_ROOT" variable specifies the root directory for the
keyfile, logfile, pidfile, and seedfile (but not the socket, which
defaults to TMPDIR since NFS can cause problems for its lockfile).
This defaults to the HOME directory for the given test.

Tests read and write files relative to their HOME / trash directory.
These files will be relocated via the use of "root" but not
"MUNGE_ROOT".  As such, replace the use of "MUNGE_ROOT" with "root"
in ".travis.yml" and "munge.spec" in order to relocate more of the
test suite cruft under "/tmp".

After diagnosing yet another Travis CI test suite failure due to
permissions, umask is now being specified in ".travis.yml" to ensure
the trash directories are created without group-writable permissions.

Tested:
- CentOS 8.2.2004
- Fedora 32