Travis: Output config.log on configure error
For Travis CI builds, output "config.log" when the configure script
fails, as was the case when trying osx builds.  This failure is
detected by grep'ing for the regex "configure: exit 1".  However,
use of the colon results in the following error in Travis CI output:

> $ {:"grep -q \"configure"=>"exit 1\" \"config.log\" && cat \"config.log\""}
> /Users/travis/.travis/functions: line 109: {:grep -q "configure=: command not found

To remedy this, escape the colon with the regex "configure[:] exit 1".

The next problem was with "config.log" output getting prematurely
truncated.  The log for each build is limited to ~4MB.  When that
limit is reached, the build is terminated and the following message is
appended to the end of the build log: "The build has been terminated."
But this was not the problem: "config.log" was well below 4MB, and the
above message was not appearing in the log.  Perhaps output was not
being flushed before the process terminated.  Adding a final "sleep",
while deeply unsatisfying, seemed to resolve the problem for now.

Reference:
- https://docs.travis-ci.com/user/common-build-problems/#log-length-exceeded