Travis: Add arm64 and ppc64le builds
Add Travis CI builds for arm64 and ppc64le CPU architectures.  This is
done by using the "jobs.include" key in order to limit expansion of
the build matrix.  Only specify builds on these architectures with
gcc and with debugging disabled for both libgcrypt and openssl.
Disable valgrind sharnanigans since these fail on ppc64le.

Note that amd64 is the default CPU architecture.

Note that jobs which are explicitly included inherit only the first
value of the defined expansion keys.  Consequently, gcc does not need
to be explicitly listed since it is the first element listed under the
"compiler" key.

Reference:
- https://docs.travis-ci.com/user/multi-cpu-architectures/
- https://docs.travis-ci.com/user/build-matrix/

Closes #79