Sharness: Fix EACCES failure succeeding for root
When the test suite is run by root, the following failure occurs in
"0103-munged-security-logfile.t":

  10 - logfile not writable by user failure

This sets the logfile perms to 0400 to check for an error when the
logfile is not writable by the user.  However, root will not get a
"permission denied" error here.  Consequently, the expected failure
erroneously succeeds.

Add a check for whether the test is being run by the root user, and
set the ROOT prerequisite when this is true.  Furthermore, add the
!ROOT prereq to the above test so it will be skipped when run by root.

Tested:
- Arch Linux
- CentOS Stream 8, 8.3.2011, 7.9.2009, 6.10
- Debian sid, 10.8, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 33, 32, 31
- FreeBSD 12.2, 11.4
- NetBSD 9.1, 9.0, 8.1
- OpenBSD 6.8, 6.7, 6.6
- openSUSE 15.2, 15.1
- Raspberry Pi OS (Raspbian 10) [armv7l]
- Ubuntu 20.10, 20.04.2 LTS, 18.04.5 LTS, 16.04.7 LTS, 14.04.6 LTS, 12.04.5 LTS