HKDF: Fix _hkdf_expand overwriting rv error status
In _hkdf_expand(), the err: label at the end of the function allows
for common cleanup before returning to the caller.  However, the
mac_cleanup() return value is assigned to "rv" which could result in
clearing a previous error condition if mac_cleanup() succeeds.

Prevent overwriting the "rv" error status by assigning to "rv2" instead.

Tested:
- Arch Linux
- CentOS Stream 8, 8.3.2011, 7.9.2009, 6.10
- Debian sid, 10.8, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 33, 32, 31
- FreeBSD 12.2, 11.4
- NetBSD 9.1, 9.0, 8.1
- OpenBSD 6.8, 6.7, 6.6
- openSUSE 15.2, 15.1
- Raspbian 10.8 [armv7l]
- Ubuntu 20.10, 20.04.2 LTS, 18.04.5 LTS, 16.04.7 LTS, 14.04.6 LTS, 12.04.5 LTS