systemd: Start at boot after network is online
At startup, munged sets its origin address.  If the "--origin="
command-line option is not specified, the origin address is determined
by calling gethostname() and gethostbyname(), the latter of which
may involve querying the network for hostname resolution.  If munged
is started at boot before the network is online, this can result in
munged setting its origin to the null address.

Delay starting munged at boot until after the network is online by
specifying "Wants=" and "After=" for "network-online.target".

Reference:
- https://www.freedesktop.org/software/systemd/man/systemd.special.html#network-online.target
- https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

Closes #93