HKDF: Revamp _hkdf_expand loop
HKDF performs a maximum of 255 rounds when expanding the pseudorandom
key to the desired length.  This commit documents that maximum and
adds HKDF_MAX_ROUNDS to replace the use of that magic number.

Each round can generate a number of bytes equal to the length of the
chosen hash function (HashLen).  hkdf() would return an error if it was
called with a dst buffer requesting more than (HashLen * 255) bytes.
But it would enter an infinite loop if the dst buffer requested 255
rounds worth of pseudorandom data.  Fortunately, neither of these
conditions could be triggered in the code calling hkdf() which is
why this was not uncovered by the sharness test suite.

The output keying material for each expansion round is computed from an
HMAC that includes a single octet denoting the given round [1..255].
This value was stored in a u_char to restrict it to a single octet,
and this variable was used as the loop control variable in the
_hkdf_expand() for-loop.  If 255 rounds were needed, the for-loop
would increment the LCV at the bottom of the loop for round 255 which
would overflow the octet back to 0 thereby causing the infinite loop.

Change the for-loop to a while-loop using "dstlen_left" in the loop
expression.  Break out of the loop if "round" equals HKDF_MAX_ROUNDS
to prevent overflow.

Additionally, remove the restriction on hkdf() that would return an
error if too large of a dst buffer was given.

Tested:
- Arch Linux
- CentOS Stream 8, 8.3.2011, 7.9.2009, 6.10
- Debian sid, 10.8, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 33, 32, 31
- FreeBSD 12.2, 11.4
- NetBSD 9.1, 9.0, 8.1
- OpenBSD 6.8, 6.7, 6.6
- openSUSE 15.2, 15.1
- Raspbian 10.8 [armv7l]
- Ubuntu 20.10, 20.04.2 LTS, 18.04.5 LTS, 16.04.7 LTS, 14.04.6 LTS, 12.04.5 LTS