Sharness: Add retry function
Add retry() in order to retry a given command up to a specified
number of times.  Moving this code into a function cleans up the
"0100-munged-lock.t" test to "start munged with leftover socket from
unclean shutdown", and makes it easy to reuse in other tests.

The loop is structured to eliminate the sleep delay whenever possible.

Tested:
- AlmaLinux 8.4
- Arch Linux
- CentOS Stream 8, 8.4.2105, 7.9.2009, 6.10
- Debian bookworm/sid, 11.0, 10.10, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Fedora 34, 33, 32
- FreeBSD 13.0, 12.2, 11.4
- NetBSD 9.2, 9.1, 9.0, 8.1
- OpenBSD 6.9, 6.8, 6.7
- openSUSE 15.3, 15.2, 15.1
- Raspbian 10.10 [armv7l]
- Ubuntu 21.04, 20.04.3, 18.04.6, 16.04.7, 14.04.6, 12.04.5