Fix unlink being interrupted by signal
On FreeBSD, unlink() can apparently be interrupted by a signal.
This behavior is not seen on any of the other platforms on which I
test, and this error is not documented in the unlink(2) manpage.

This failure was found by "check pidfile removal" in "0010-basic.t".
The multiple SIGTERMs sent to stop the daemon would occasionally
interrupt the unlink() of the pidfile in destroy_conf().
Unfortunately, tracking down the source of the failure was complicated
by the return value of this unlink() being discarded instead of used
to log a warning when an error occurred.  (This is why you should
always check return values!)

Fix this potential error by placing unlink()s in a do-while loop to
retry EINTR errors.  Additionally, always check the return value and
log errors.

Note that the legacy code in "src/libmunge/auth_send.c" and
"src/munged/auth_recv.c" is not being changed since it is difficult
to test for correctness.

Tested:
- AlmaLinux 8.4
- Arch Linux
- CentOS Stream 8, 8.4.2105, 7.9.2009, 6.10
- Debian bookworm/sid, 11.0, 10.10, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Fedora 34, 33, 32
- FreeBSD 13.0, 12.2, 11.4
- NetBSD 9.2, 9.1, 9.0, 8.1
- OpenBSD 6.9, 6.8, 6.7
- openSUSE 15.3, 15.2, 15.1
- Raspbian 10.10 [armv7l]
- Ubuntu 21.04, 20.04.3, 18.04.6, 16.04.7, 14.04.6, 12.04.5

The failure in "0010-basic.t" was reproduced on FreeBSD 12.2 and
11.4 before the changes.  Afterwards, this check was successfully
run for over 100k iterations on both 12.2 and 11.4.  Consequently,
it is being changed from test_expect_unstable to test_expect_success.