Job: Throttle log msg for failed accept
When munged accepts a new connection, accept() errors for EMFILE,
ENFILE, ENOBUFS, and ENOMEM cause the accept() loop to block while
the backlog is processed.  But the resulting log message does not
differentiate between these errors, and repeated messages are not
rate-limited.  This can result in rapid growth of the logfile if
the system has reached the open file descriptor limit or exhausted
system memory.  Furthermore, the client will retry failed requests
up to 5 times, thereby increasing these messages when the system is
in this depleted state.

Throttle these log messages for failed accept()s to one per minute.
This seems like a good compromise between recording errors and limiting
the chance of filling the filesystem with redundant messages.

Additionally, include strerror() in the log message in order to
differentiate between possible errors.

Tested:
- Arch Linux
- CentOS Stream 8, 8.3.2011, 7.9.2009, 6.10
- Debian sid, 10.8, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 33, 32, 31
- FreeBSD 12.2, 11.4
- NetBSD 9.1, 9.0, 8.1
- OpenBSD 6.8, 6.7, 6.6
- openSUSE 15.2, 15.1
- Raspbian 10.8 [armv7l]
- Ubuntu 20.10, 20.04.2, 18.04.5, 16.04.7, 14.04.6, 12.04.5

There isn't a good way to automatically test this in sharness without
mucking up the code, so the accept() failure case was manually tested
on Debian 9.13.

Closes #94