libmunge: Add suggestion to err msg to start munged
Trying to connect to the munge socket when munged is not running
results in a potentially confusing error message:

  munge: Error: Failed to access "/run/munge/munge.socket.2":
    No such file or directory

This error message is correct in that the client was unable to access
the socket because it did not exist (ENOENT).  But the user might
not understand what that means or how to correct it.

Add a suggestion to the error message to start munged.

Tested:
- Debian 9.13