munged: Add suggestion to err msg to run mungekey
Starting munged without having created a key beforehand results in
a potentially confusing error message:

  munged: Error: Failed to check keyfile "/etc/munge/munge.key":
    No such file or directory

This error message is correct in that the keyfile could not be checked
because it does not exist (ENOENT).  But the user might not understand
how to correct it.

Add a suggestion to the error message to run mungekey.

Tested:
- Debian 9.13