munged: Update --stop err msg for process not running
Attempting to stop a munged that is not running results in the
following error message (only if the --verbose option is specified):

  munged: Error: Failed to query socket lockfile
    "/run/munge/munge.socket.2.lock": No such file or directory

This error message is correct in that the socket lockfile does
not exist.  The lockfile is used to check the advisory lock held by
the running munged process from which the pid can be ascertained.
If the lockfile does not exist, the munged process listening on the
corresponding socket is presumably no longer running.  But the user
might not understand the implication of this message.

Add a note to the error message that a running process could not
be found.

Tested:
- Debian 9.13