Sharness: Set IFNAME prereq if network ifname found
The Debian kfreebsd-i386 build for 0.5.14-4 encountered the following
failures in "0110-munged-origin-addr.t":

  10 - munged --origin interface name
  11 - munged --origin interface name metadata

The test log output showed the following message from the check "munged
--origin local IP address" that uses the "--origin=127.0.0.1" option
and saves the loopback network interface name to file for later checks:

  2021-02-28 01:09:43 +0000 Info:      Set origin address to 127.0.0.1 ()
  ok 8 - munged --origin local IP address
  PASS: 0110-munged-origin-addr.t 8 - munged --origin local IP address

Instead of the empty parentheses here, I would expect to see the string
"lo0" associated with 127.0.0.1 on Debian GNU/kFreeBSD.

munged/conf.c:write_origin_addr() logs this message.  But the
trailing "(%s)" string containing the network interface name is
only added if conf->origin_ifname != NULL.  That variable is set by
munged/net.c:_net_get_hostaddr_via_ifaddrs() if a network interface
matching the specified origin is found.  This implies a matching
interface is being found and the origin ifname string is being set
to the empty string (or some sequence of non-printable characters).
And that seems strange and somewhat unlikely.

As mentioned above, this sharness check specifies "--origin=127.0.0.1",
greps the log for the expected message, and saves the interface name
appearing within the parentheses to the file "ifname0.$$".  This file
is read by the two failing checks above to test whether an origin
address can be specified via network name.  Performing the check in
this manner seems the best alternative since there are no portable
means to enumerate the network interfaces on the system.

Based on the test log output, it's unclear exactly why these tests
are failing.  It appears that a matching interface name is found
due to the parentheses.  But if so, the sed expression run against
empty parentheses would create the file "ifname0.$$" containing only
a newline, which would cause "test -s" to evaluate true since the
filesize is non-zero, which would start munged with --origin="", which
would write the message 'Error: Failed to lookup origin ""' to stderr.
However, that message does not appear in the test log output.

Another possibility is that a matching interface name is not found,
which would result in "ifname0.$$" being empty, which would cause
"test -s" to evaluate false, which would cause the check to immediately
fail without any output.  But that doesn't explain the log message
with the empty parentheses.

I'm unable to reproduce these failures on a freshly-installed Debian
kfreebsd-i386 vm; these tests succeed.  On my vm, the lo0 network
interface has the IP address 127.0.0.1 as expected.

This commit makes the following changes:

  1. Remove the "test -s ifname0.$$" statements from the sharness
  checks.  They make it difficult to diagnose why a check fails.
  If the file is empty, munged will fail with the error 'Failed to
  lookup origin ""' which provides useful information for debugging.

  2. Create the new check 'munged --origin interface name lookup' which
  greps the log from the preceding check for the loopback interface
  name, checks that it is not the empty string, saves the name to
  the file "ifname0.$$", and sets the sharness IFNAME prerequisite.

  3. Change the checks for 'munged --origin interface name' and
  'munged --origin interface name metadata' to depend on this new
  IFNAME prerequisite.  Thus, if munged is unable to match 127.0.0.1 to
  an interface name, these checks will be skipped instead of failing.

  4. Change _net_get_hostaddr_via_ifaddrs() to check that ifa_name
  is not the empty string before assigning the string for the network
  interface.

Tested:
- Arch Linux
- CentOS Stream 8, 8.3.2011, 7.9.2009, 6.10
- Debian sid, 10.8, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Debian GNU/kFreeBSD sid [amd64, i386]
- Fedora 33, 32, 31
- FreeBSD 12.2, 11.4
- NetBSD 9.1, 9.0, 8.1
- OpenBSD 6.8, 6.7, 6.6
- openSUSE 15.2, 15.1
- Raspbian 10.8 [armv7l]
- Ubuntu 20.10, 20.04.2, 18.04.5, 16.04.7, 14.04.6, 12.04.5

Further tested by manually specifying "--origin=127.0.0.2" instead in
"0110-munged-origin-addr.t" to simulate failing to match the loopback
IP address to an interface name.