Build: Fix makefile uninstall-systemd prereq
When running "make uninstall" on a systemd-based system, the sysconfig
file is being uninstalled by "uninstall-sysvinit-sysconfig" instead of
"uninstall-systemd-sysconfig".  The end result is the same, but this
will cause problems when additional systemd files are installed.

Add the missing "uninstall-systemd-sysconfig" prerequisite to the
"uninstall-systemd" rule.  Additionally, update the test conditions for
both "uninstall-systemd-sysconfig" and "uninstall-sysvinit-sysconfig"
so they match their corresponding install rule test conditions.
This will result in "uninstall-systemd-sysconfig" correctly removing
the sysconfig file.