munged: Remove repeated SIGTERMs for signaling stop
In _conf_process_stop(), sending multiple SIGTERMs is not necessary;
in fact, it may be harmful since these unnecessary signals can
interrupt calls on some platforms resulting in dropped log messages
or an unclean shutdown where transient files are not removed.

For the new "stop" behavior, send a single SIGTERM and wait up to
MUNGE_SIGNAL_WAIT_MSECS milliseconds for the process to terminate.
If necessary, send a subsequent SIGKILL and wait up to an additional
MUNGE_SIGNAL_WAIT_MSECS milliseconds for the process to terminate.

In _conf_send_signal(), send the specified signal and check for
process termination every MUNGE_SIGNAL_CHECK_MSECS milliseconds up
to the specified maximum delay.

Additionally, add clock_is_timespec_expired().

Tested:
- AlmaLinux 8.4
- Arch Linux
- CentOS Stream 8, 8.4.2105, 7.9.2009
- Debian bookworm/sid, 11.0, 10.10, 9.13, 8.11
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Fedora 34, 33, 32
- FreeBSD 13.0, 12.2, 11.4
- NetBSD 9.2, 9.1, 9.0, 8.1
- openSUSE 15.3, 15.2, 15.1
- Raspbian 10.10 [armv7l]
- Ubuntu 21.04, 20.04.3, 18.04.6, 16.04.7, 14.04.6

Failed:
- CentOS 6.10
- Debian 7.11, 6.0.10, 5.0.10, 4.0
- Ubuntu 12.04.5
  munged-clock.o: In function `clock_get_timespec':
  src/munged/clock.c:51: undefined reference to `clock_gettime'
  munged-conf.o: In function `_conf_send_signal':
  src/munged/conf.c:841: undefined reference to `clock_nanosleep'
- OpenBSD 6.9, 6.8, 6.7
  src/munged/conf.c: In function '_conf_send_signal':
  src/munged/conf.c:841: warning: implicit declaration of function 'clock_nanosleep'

The occasional 'check logfile for stop' failure in "0010-basic.t"
occurring on FreeBSD appears to be fixed by this commit.  It was
successfully run for over 100k iterations on 13.0, 12.2, and 11.4.
Consequently, it is being changed from test_expect_unstable to
test_expect_success.