Remove GCRYPT_VERSION from gcry_check_version
Debian reports munged is incorrectly checking the minimum required
version of Libgcrypt.  This is causing munged to fail on an update
from buster to bullseye since it was built with a GCRYPT_VERSION of
1.8.8 but is being installed with libgcrypt20 1.8.7-6.

According to the Libgcrypt documentation, gcry_check_version()
should be called with the minimum required version of the library
(or NULL if that check is not needed).  The <gcrypt.h> header file
further notes GCRYPT_VERSION should not be used by the program since
gcry_check_version() should return the same version string.

Replace GCRYPT_VERSION with NULL in gcry_check_version() to disable the
version check.  Debian further notes their automated system determined
v0.5.14 requires a minimum Libgcrypt version of 1.8.0.  However, it
seems preferable to disable the check and let distributions perform
their dependency checks instead of maintaining this value by hand.
Note that gcry_check_version() must still be called because it also
initializes the library.

Reference:
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=991875
- https://gnupg.org/documentation/manuals/gcrypt/Initializing-the-library.html

Tested:
- AlmaLinux 8.4
- Arch Linux
- CentOS Stream 8, 8.4.2105, 7.9.2009, 6.10
- Debian sid, 10.10, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Debian GNU/kFreeBSD sid [amd64, i386]
- Fedora 34, 33, 32
- FreeBSD 13.0, 12.2, 11.4
- NetBSD 9.2, 9.1, 9.0, 8.1
- OpenBSD 6.9, 6.8, 6.7
- openSUSE 15.2, 15.1
- Raspbian 10.10 [armv7l]
- Ubuntu 21.04, 20.04.2, 18.04.5, 16.04.7, 14.04.6, 12.04.5