Sharness: Expect success when checking logfile for replay
Update the sharness test checking the logfile for a message from
an earlier test that replayed a credential.  Change it from being
unstable to expect success.

Additionally, for clarity, move this logfile check to immediately
follow the replayed credential check that generates the log message.
This undoes the reordering of tests in commit baf36a3 that attempted
to remedy an occasional NetBSD failure.

"0010-basic.t" was successfully run for over 100k iterations on
NetBSD 9.2, 9.1, and 8.1.  But on NetBSD 9.0, the following failure
was observed after 8094 iterations:

> not ok 4 - start munged
> #
> #           munged_start_daemon
> #
> munged: 2021-10-13 04:31:21 +0000 Error: Failed to validate lockfile:
>   "/tmp/munged.sock.27439.lock" should only have permissions for write
>   by user (file mode 0200)

On that iteration, the lockfile was curiously permissioned 0222.

Tested:
- AlmaLinux 8.4
- Arch Linux
- CentOS Stream 8, 8.4.2105, 7.9.2009, 6.10
- Debian bookworm/sid, 11.1, 10.11, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Fedora 34, 33, 32
- FreeBSD 13.0, 12.2, 11.4
- NetBSD 9.2, 9.1, 9.0, 8.1
- OpenBSD 6.9, 6.8, 6.7
- openSUSE 15.3, 15.2, 15.1
- Raspbian 10.10 [armv7l]
- Ubuntu 21.04, 20.04.3, 18.04.6, 16.04.7, 14.04.6, 12.04.5