Sharness: Add --verbose to munged_stop_daemon
Add the "--verbose" option to the "munged --stop" command in
munged_stop_daemon().  This allows the caller to grep stderr in
order to determine whether the process gracefully terminated or was
forcibly killed.

Additionally, update instances of munged_stop_daemon() in
"0100-munged-lock.t" where a graceful shutdown is necessary for the
subsequent check to succeed.

Tested:
- AlmaLinux 8.4
- Arch Linux
- CentOS Stream 8, 8.4.2105, 7.9.2009, 6.10
- Debian bookworm/sid, 11.1, 10.11, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Fedora 34, 33, 32
- FreeBSD 13.0, 12.2, 11.4
- NetBSD 9.2, 9.1, 9.0, 8.1
- OpenBSD 6.9, 6.8, 6.7
- openSUSE 15.3, 15.2, 15.1
- Raspbian 10.11 [armv7l]
- Ubuntu 21.04, 20.04.3, 18.04.6, 16.04.7, 14.04.6, 12.04.5