Cleanup: md: Remove magic and finalized
Remove the 'magic' and 'finalized' variables used in the debug build.

The 'magic' variable was used to detect memory corruption that
overwrote part of the md_ctx struct.  There are better ways to detect
memory corruption nowadays.

The 'finalized' variable was used to ensure the md_ctx was not updated
after having been finalized, or finalized more than once.  It also
ensured for md_copy() that the src md had been finalized and the dst
md had not.  Its use required the dubious practice of performing an
assignment within an assert().