Cleanup: cipher: Replace parm-validating asserts
Replace assert()s used for parameter validation at the start of
extern functions with if-conditions that are always checked regardless
of NDEBUG.  'dstlen' is dereferenced and checked to ensure it is not
negative since the corresponding static functions do not check.

Remove assert()s used for parameter validation at the start of static
(internal) functions since they add to clutter and these parameters
have already been validated by the corresponding extern functions.