Update TESTS with stubs to include in release tar
The sharness test files "0104-munged-security-pidfile.t"
and "0105-munged-security-seedfile.t" are stubs added back in
commit 49f0b7a still awaiting future development.  But these files are
not distributed with the release tarball since they are not specified
in "t/Makefile.am" for either TESTS or EXTRA_DIST.  This omission
seems to have confused Fedora which has added them to their specfile
(even though they are not run during a check).

To avoid further confusion downstream, add these two files to TESTS
in order for them to be distributed with the release tarball and run
during a check (even though no meaningful checks are yet performed).