Sharness: Fix 0101-munged-security-socket.t cleanup
munged_cleanup() is called at the end of an &&-chain in the final
check.  As such, it can be prevented from running by a failure of
a preceding command in the &&-chain.  In fact, rmdir will fail if
munged is still running with its socket residing in the directory
that the cleanup test is trying to remove:

> /usr/bin/rmdir: failed to remove '/tmp/munge-31415/socketdir-31415':
>   Directory not empty

Move munged_cleanup() to the start of the &&-chain to ensure it will
not be prevented from running due to a preceding failure in the chain.

Also note that munged_cleanup() must always be the first command when
it appears in an &&-chain.