Build: Fix gen-date & gen-version for git-worktree
When working in a git worktree, "gen-date" and "gen-version" do not
set the expected date and version strings according to git metadata.
This was fortuitously detected by the recent change to "bootstrap"
to output version info after the configure script had been updated.

These build-aux scripts were introduced in commit 1a6c9e0 and check
for the presence of a git checkout via "test -d ${GIT_DIR}".  However,
a new git worktree (i.e., "linked worktree") links to the ${GIT_DIR}
in the "main worktree".  As such, it is a regular file and not
a directory.

Fix the check for the presence of a git checkout by instead using
"test -e ${GIT_DIR}".