Sharness: Fix regex in -V/--version check
The existing regex pattern can fail if the version information
is empty.  As such, it will accept any string containing "munge-".

Fix the grep regex pattern for checking "--version" to require the
version string start with a digit, contain only hexadecimal digits
and periods, and end with a space character delimiter.  Note that
hexadecimal digits are needed here since snapshot release versions
are appended with the hash of the latest git commit.

Tested:
- AlmaLinux 9.0, 8.6
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 36, 35, 34
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.1, 7.0, 6.9
- openSUSE 15.4, 15.3, 15.2, 15.1
- Raspbian GNU/Linux 10.12 [armv7l]
- Ubuntu 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5