Sharness: Replace regex symbols \< and \>
The symbols \< and \> respectively match the empty string at the
beginning and end of a word.  However on OpenBSD, these are provided to
ease compatibility with traditional SVR4 systems but are not portable
and should be avoided.

Replace the use of \< and \> with either a space character (if
sufficient), or an extended regex to match the expected word pattern
immediately followed by either a space character or end of line.

Also add a "0002-smiple.t" check to verify this replacement is working
as expected.

Reference:
- re_format(7) on OpenBSD 7.1

Tested:
- AlmaLinux 9.0, 8.6
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 36, 35, 34
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.1, 7.0, 6.9
- openSUSE 15.4, 15.3, 15.2, 15.1
- Raspbian GNU/Linux 10.12 [armv7l]
- Ubuntu 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5