Sharness: Change mungekey h/L/V checks to use pipe
Change the mungekey checks for "--help", "--license", and "--version"
to have grep read from a pipeline instead of a temporary file.  This is
safe since the exit status of a pipeline is the exit status of the
last command in the pipeline, and grep can only succeed here if the
mungekey executable outputs the correct text onto stdout.  This also
mirrors how these options are checked in "0011-munged-cmdline.t",
"0012-munge-cmdline.t", and "0013-unmunge-cmdline.t".

Tested:
- AlmaLinux 9.0, 8.6
- Arch Linux
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009, 6.10
- Debian GNU/kFreeBSD bookworm/sid [amd64, i386]
- Debian GNU/Linux sid, 11.5, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 36, 35, 34
- FreeBSD 13.1, 13.0, 12.3
- NetBSD 9.3, 9.2, 9.1, 9.0
- OpenBSD 7.1, 7.0, 6.9
- openSUSE 15.4, 15.3, 15.2, 15.1
- Raspbian GNU/Linux 10.12 [armv7l]
- Ubuntu 22.04.1, 20.04.5, 18.04.6, 16.04.7, 14.04.6, 12.04.5