Sharness: Fail &&-chain if no builddep command
In "1000-chaos-rpm.t", the builddep command check assumes either "dnf"
or "yum-builddep" is installed.  If neither is found, BUILDDEP will
be unassigned and the resulting sudo command will try to exec an SRPM
to no avail.

Fail the sharness &&-chain if a builddep command is not found.
Additionally, redirect the output from the "type" builtin to /dev/null.

Tested:
- AlmaLinux 9.0, 8.6
- CentOS Linux Stream 9, Stream 8, 8.5.2111, 7.9.2009
- Fedora 36, 35, 34