Spec: Remove alternate source file verification
Fedora Packaging Guidelines state that packages should verify
OpenPGP signatures (if available) as part of the RPM build process.
Fedora originated the "%{gpgverify}" rpm macro to facilitate this
source file verification.  However, RHEL did not provide this macro,
so the rpm spec file added an alternate implementation for when
"%{gpgverify}" was not defined.

RHEL now provides this macro via updates to the redhat-rpm-config
package in order to better unify Fedora and RHEL spec files.
Correspondingly, this macro now exists on recent versions of AlmaLinux
and CentOS.  Since the alternate implementation is no longer needed,
remove it from the spec file.

Reference:
- https://docs.fedoraproject.org/en-US/packaging-guidelines/#_source_file_verification
- https://fedoraproject.org/wiki/PackagingDrafts:GPGSignatures
- https://rpm-software-management.github.io/rpm/manual/conditionalbuilds.html
- https://src.fedoraproject.org/rpms/redhat-rpm-config/pull-request/54
- https://bugzilla.redhat.com/show_bug.cgi?id=1874576

Tested:
- AlmaLinux 9.0 (redhat-rpm-config-194-1.el9.alma.noarch)
- AlmaLinux 8.6 (redhat-rpm-config-129-1.el8.alma.noarch)
- CentOS Stream 9 (redhat-rpm-config-196-1.el9.noarch)
- CentOS Stream 8 (redhat-rpm-config-130-1.el8.noarch)
- CentOS Linux 8.5.2111 (redhat-rpm-config-125-1.el8.noarch)
- CentOS Linux 7.9.2009 (redhat-rpm-config-9.1.0-88.el7.centos.noarch)
- Fedora 36, 35, 34